requirements.txt
Pillow==9.4.0
pypng==0.20220715.0
PyQRCode==1.2.1
qrcode==7.4.2
typing_extensions==4.5.0
